+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Refactor Softmaxpro adapter to use the new schema mapper.
+- Rename Nanodrop Eight adapter to Nanodrop 8000
 
 ### Deprecated